refactor(extension): modularize progress view and command helpers
Split latexdiff commands, progress-view formatters, and stream tab
components into focused modules behind unchanged entry points.
